fcontext make_fcontext param use size_t replace unsigned type

+++ fcontext/src/fcontext.h
@@ -6,6 +6,7 @@
 
 #ifndef BOOST_CONTEXT_DETAIL_FCONTEXT_H
 #define BOOST_CONTEXT_DETAIL_FCONTEXT_H
+#include <stddef.h>
 #ifdef  __cplusplus
 extern "C" {
 #endif
@@ -25,7 +26,7 @@ typedef struct {
     void* data;
 } transfer_t;
 transfer_t BOOST_CONTEXT_CALLDECL jump_fcontext(fcontext_t const to, void* vp);
-fcontext_t BOOST_CONTEXT_CALLDECL make_fcontext(void* sp, unsigned size, void (*fn)(transfer_t));
+fcontext_t BOOST_CONTEXT_CALLDECL make_fcontext(void* sp, size_t size, void (*fn)(transfer_t));
 transfer_t BOOST_CONTEXT_CALLDECL ontop_fcontext(fcontext_t const to, void* vp, transfer_t(*fn)(transfer_t));
 
 #ifdef  __cplusplus
